Rust binding multithreading
Since the C++ index is thread-safe, we should expect it to be behind a
shared pointer which we can clone and move across threads safely in
rust. This allows naive copying and &self calling in a high concurrency
runtime like tokio.

I've encountered errors when incurring concurrency on the reserve()
function, which indicates that while the main index might be thread
safe, some of the metadata might not be well protected. Other than that,
I have only successfully tested concurrent calls to the add method.